:arrow_up: chore(website): update website submodule to latest commit on develop branch (#6)
* 🐛 fix(package.json): replace bvm install with git submodule update
The bvm install command is no longer needed as the required dependencies are now included as a git submodule. The new command git submodule update --init --recursive will ensure that the submodule is up to date and initialized.

* 🔀 chore(.gitmodules): add develop branch to website submodule
🔀 chore(website): update website submodule to latest commit on develop branch
The website submodule has been updated to the latest commit on the develop branch. The .gitmodules file has been updated to include the develop branch for the website submodule. This change allows the website submodule to be updated to the latest changes on the develop branch.