feat: add README.md file (#1)
The README.md file has been added to the profile directory. It contains an introduction to the Eventiva platform, an index, information on how to get support, and instructions on how to contribute to the project. The README.md file is intended to provide a quick overview of the project and help new contributors get started.