📝 docs: mono-1 update documentation for org change  (#9)
* 🎉 feat(compass.yml): add compass.yml file
The compass.yml file is added to the project. This file contains metadata about the project, including its name, description, version, and links to relevant resources. It also includes information about the project's license and contributors. This file is used by Atlassian's Compass tool to manage the project's metadata.

* 🗑️ chore(github): remove unused GitHub workflows - MONO-3
The three GitHub workflows, cla.yml, create-diagram.yml, and github_management.yml, are no longer used and have been removed from the repository.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* 📝 docs(all): update project name and owner, and fix links in GitHub files - MONO-1
This commit updates the project name and owner in the .all-contributorsrc file and the GitHub files. It also fixes the links in the files to point to the correct repository. The changes were made to ensure consistency and accuracy in the project's documentation.

📝 docs: update project name and links in issue templates
The project name and links in the issue templates have been updated to reflect the new project name and repository location. This ensures that all references to the project are up-to-date and accurate.

🚀 chore(config): update project URLs and author information
This commit updates the URLs for the project schema, dependabot configuration, git submodules, and README badges to reflect the new repository location under the Eventiva organization. It also updates the author information in the dependabot and README files to reflect the new email domain and organization name.

🚚 chore(preroot.js): update project name and links
📝 docs(contributing.md): update links to contributors folder and project name
🚀 chore(package.json): update project name and repository link
🚀 chore(rush.json): update repository link
The project name and repository links have been updated to reflect the new name of the project, which is now @eventiva/eventiva. The links in the contributing.md file have also been updated to reflect the new name and location of the contributors folder.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* 🔧 chore(settings): update author email and add company name
The author email has been updated to the new email address. The company name has been added to the settings file to improve consistency across the project.

* 🔀 chore(.gitmodules): add submodules for .github and contributors
Three submodules have been added to the repository: .github, contributors, and website. The .github and contributors submodules have been added to the projects directory. The .github submodule is a repository containing GitHub templates and workflows, while the contributors submodule is a repository containing a list of contributors to the project.

---------

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>