✨chore(workflows): update workflows (#2)
* 🚚 refactor(diagram): update file path and project name

This commit updates the file path for create-diagram.yml and project name to @eventiva/eventiva. It also modifies the "push" event to the "pull_request_target" event with specific types and ignored branches to match the workflow. Lastly, it updates the "Last Modified" and "Author" fields.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* ✨ feat(infra): add GitGuardian iac scan workflow

Adds a new workflow infrastructure_scan.yml to scan infrastructure as code (IaC) using GitGuardian. The workflow runs on every push and pull request. It uses the GitGuardian/ggshield/actions/iac@main action and checks minimum-severity "CRITICAL". The workflow checks the ./terraform/ directory and uses the secret GITGUARDIAN_API_KEY.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* ✨ feat(secrets.yml): Add GitGuardian scan for push and pull_request

This commit adds a new Github Action, GitGuardian scan, that runs on every push and pull_request. This Action scans multiple commits to check for leaked secrets, unauthorized access, and policy violations.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* :recycle: chore(infrastructure_scan.yml): improve filtering logic for GitGuardian scan job

This commit adds path filtering for the `GitGuardian iac scan` job in the workflow, making it only trigger when the `.terraform` directory is present and modified. This filter is now implemented for both push and pull-request events.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

---------

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>