chore: workflow fixes (#10)
* 🚀 chore(devcontainer.json): add ggshield to postCreateCommand

This change adds a command to the devcontainer's postCreateCommand property to install and authenticate ggshield. This will be used for identifying secrets in our GitHub repositories in the future.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* :sparkles: feat(pre-commit): add ggshield secret scan

This commit adds the ggshield secret scan command to the pre-commit git hook. The new command scans the repository for secrets and will fail the commit if any are found.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* 🔀 chore(projects): update subproject commit

The subproject commit has been updated to 322ba432a0ac240543c625cbded3000806919991.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* 🔀 chore(rush): update rush/pnpm lock

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* 🔄 chore(.gitmodules): update github repository URL

Change the URL of the website github repository from eventiva-website to Directory.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* :recycle: refactor(.gitmodules): remove unnecessary submodule and fix spacing

Removed an unnecessary submodule "projects/website" from .gitmodules, and fixed the spacing to conform to the expected format.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* :sparkles: feat(rush.json): add project @eventiva/website to prototypes

Added @eventiva/website as a new prototype project to rush.json's projects array.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

---------

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>