feat: added bug report and config templates (#5)
* :sparkles: feat(issue templates): added bug report and config templates

This commit adds two new issue templates, bug_report.yml and config.yml. The bug_report.yml template provides a structure for users to report bugs, including information about the affected features and a section for reproducing the issue. The config.yml template is a basic template for requesting configuration changes. These templates will help streamline the process of reporting bugs and requesting configurations.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

* :recycle: fix(github_management.yml): Update project and file paths

- Updated the project name to @eventiva/eventiva.
- Updated the file path to \projects\.github\.github\workflows\github_management.yml.
- Updated the contributing guidelines and code of conduct URLs to reflect the new project name @eventiva/eventiva.
- Updated the copyright holder to Eventiva.
- Updated the last modified date.
- Updated the email address to jonathan.stevens@eventiva.co.uk.

This change ensures that all references to the project and file paths are up to date and aligned with the current state of the @eventiva/eventiva project.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>

---------

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>