chore: create workflow to post change report (#22)
🚧:construction: chore(discord_changelog.yml): create workflow to post change report

This commit adds a new workflow file, discord_changelog.yml, which sets up a scheduled job to post a change report to Discord. The change report includes information about the changes made in the past day. The report is generated using the maxprilutskiy/change-report action and requires an OpenAI API key and a Discord bot token.

The purpose of this workflow is to keep the Discord community updated on the latest changes in the project.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>