feat: add new DocType
This commit adds a new DocType called "Discord Tag" to the "Discord Integration" module. It includes fields for name, emote, and an option to restrict tag application to moderators only.

The addition of this new DocType enhances the functionality of the Discord Integration module by allowing users to define and manage tags for Discord integration within the system.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>