🔀 chore(license): update license file
The license file has been updated from Creative Commons to GNU General Public License (GPL) Version 3. This change ensures that the software and other works are protected under the copyleft license. The new license guarantees the freedom to share, modify, and distribute the software while also respecting the rights of the authors and developers. This update aligns with the project's commitment to open-source principles.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>