feat: Add workflow for committing bitmap
This commit adds a new workflow file "commit-bitmap.yml" that sets up a workflow to commit the bitmap of the repository. The workflow is triggered manually and runs on the latest version of Ubuntu. It initializes Bit, tags and exports, and then commits the bitmap. This workflow will help in keeping track of the changes made to the repository and create a historical record of the modifications.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>