feat: Add Tag and Export workflow
This commit adds a new GitHub Actions workflow file called `tag-export.yml` that sets up a workflow to tag and export the repository using Bit. It runs on the latest version of Ubuntu and includes environment variables for authentication.

The workflow performs the following steps:
1. Checkout the repository
2. Initialize Bit
3. Tag and export using Bit

This addition will make it easier to manage and automate the tagging and exporting process for the repository.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>