feat: add git pull command
Add the "git pull --recurse-submodules" command to the initialization process in the devcontainer.json file. This command ensures that the repository and its submodules are all up to date when the devcontainer is created. The addition of this command makes it easier for developers to work with the latest version of the codebase.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>