:fire: Remove Directory submodule and update URL for vetting submodule
Previously, the "Directory" submodule was included in the `.gitmodules` file, but it has been deleted in this commit. Additionally, the URL for the "vetting" submodule has been updated to `git@github.com:Eventiva/vetting.git`. This change improves the organization of the project and ensures that the correct submodule URL is being used.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>