feat: add custom fields for course chapters, course lessons, and LMS courses
This commit adds custom fields for "Company" in the "Course Chapter," "Course Lesson," and "LMS Course" doctypes. The custom fields are of type "Link" and will be visible in list views and standard filters. These fields allow users to associate a company with the respective courses. The custom fields are non-mandatory and have no predefined default values.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>