feat: add Sweep label
This commit adds a new label called "sweep" in the .github/config.json file. The "sweep" label has been created for issues and pull requests that are being handled by Sweep, an AI-powered junior developer. The label has a color code of CFD3D7. This change enhances the organization and categorization of tasks within the project.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>