chore: add Microsoft Defender For DevOps workflow
This commit adds a new workflow file, `defender-for-devops.yml`, which sets up the integration with Microsoft Defender For DevOps. The workflow is triggered on pushes and pull requests to the `develop` branch, as well as on a weekly schedule. It runs on the `windows-latest` runner.

The workflow consists of the following steps:
1. Checkout the repository
2. Setup .NET environment with versions 5.0.x and 6.0.x
3. Run Microsoft Security DevOps action using the `microsoft/security-devops-action@v1.6.0` action
4. Upload the results to the Security tab using the `github/codeql-action/upload-sarif@v2` action

Please note that this workflow requires a separate integration with Microsoft Defender For DevOps and appropriate permissions to report data back to Azure. For more information, refer to the official documentation: [https://learn.microsoft.com/en-us/azure/defender-for-cloud/quickstart-onboard-github](https://learn.microsoft.com/en-us/azure/defender-for-cloud/quickstart-onboard-github)