:arrow_up: chore(bit-init): Bump bit version
Update `bit-init.js` and `bit-init.sh` files to set the `BIT_VERSION` variable to an empty string. This allows the script to install the latest version of Bit. This change is necessary to ensure compatibility with the latest features and bug fixes in Bit.

By removing the specific version number, we can leverage the `@teambit/bvm` package to install the latest version of Bit automatically. This simplifies the installation process and ensures that users always have the most up-to-date version.

This commit adheres to Semantic Versioning standards, as it does not introduce any new features or backward-incompatible changes but rather updates the configuration file to support the latest version of Bit.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>