🔥 chore(test-more-bash): remove .gitrepo file
The .gitrepo file was deleted as it is no longer needed for maintaining the git "subrepo" in the test-more-bash directory. This file was previously used by the git-subrepo command to track the remote, branch, commit, parent, cmdver, and method information. However, it is no longer necessary and has been removed.