🔥 chore(bashplus): remove .gitrepo file
The .gitrepo file was deleted as it is no longer needed. This file was used to maintain the git "subrepo" configuration for the bashplus subdirectory. The subrepo was pointing to the git repository at git@github.com:ingydotnet/bashplus.git with the branch set to master and a specific commit (e49f45a1457fed3cceb15bd4a82b0f7515efd8e5). The parent commit was c978e2afd6861203138f28d0021e03fa1ffbba0c and the cmdver was 0.4.1. The method used for merging was merge.