🔥 chore(test-tap-bash): remove .gitrepo file
The .gitrepo file was deleted as it is no longer needed. This file was used to maintain the git "subrepo" configuration for the test-tap-bash project. The subrepo was pointing to the remote repository at git@github.com:ingydotnet/test-tap-bash.git, branch master, commit a9b33446848440a445f664a942f676d726788eff. However, this subrepo is no longer being used, so the .gitrepo file was removed.