📝 docs(compass.yml): add compass.yml file for the mono-repository component
This commit adds the `.devcontainer/compass.yml` file to the project. The `compass.yml` file provides metadata and configuration information for the mono-repository component. It includes details such as the project name, author information, contributing guidelines, code of conduct, and licensing information.

The `compass.yml` file is used to provide context and documentation for the mono-repository component, making it easier for developers to understand and contribute to the project.

For more information on the formatting of `compass.yml` files, please refer to the [Compass YML Format documentation](https://go.atlassian.com/compass-yml-format).