docs: add compass.yml file for Github Workflows in the projects/workflows/.github directory
This commit adds the compass.yml file to the projects/workflows/.github directory. The compass.yml file contains the configuration for Github Actions and Rule Sets for the Organisation. It includes information such as the project name, description, config version, type ID, owner ID, and other fields.

The compass.yml file is used to define the workflows and actions that will be executed in the Github repository. It provides a way to automate various tasks and processes in the development workflow.

For more information on the formatting of the compass.yml file, please refer to the documentation at https://go.atlassian.com/compass-yml-format.