chore: add chore-trunk-nightly.yml workflow file
This commit adds the `.github/workflows/chore-trunk-nightly.yml` file, which contains the configuration for the "Chore - Trunk Nightly" workflow. This workflow is triggered on a schedule and also manually through workflow dispatch.

The workflow consists of two jobs: `trunk_upgrade` and `trunk_check`. The `trunk_upgrade` job runs on an Ubuntu environment and performs an upgrade using the Trunk tool. It has write permissions for contents and pull requests. The `trunk_check` job also runs on an Ubuntu environment and performs a check using the Trunk tool. It has write permissions for checks and read permissions for contents.

Both jobs include steps to harden the runner, collect workflow telemetry, checkout the repository, and perform other necessary actions specific to the Trunk tool.

This workflow is part of the Eventiva project and is licensed under the GNU General Public License v3.0.