chore: Remove outdated issue templates
The commit removes a series of issue templates from the .github/ISSUE_TEMPLATE directory, including bug reports, maintenance requests, doc reviews, fixes, and a template for miscellaneous issues. This housekeeping effort might be in response to a restructuring of the project's issue submission process or a move to a different issue tracking methodology.

Signed-off-by: Jonathan Stevens <jonathan.stevens@eventiva.co.uk>